tc-5083: rename tags to labels, store tag to tag
