+++ css/scss/apps/stores/_modal.scss
@@ -9,7 +9,7 @@
   }
 }
 
-.tags-modal {
+.labels-modal {
   max-height: 20em;
   overflow-y: scroll;
 }
