+++ js/apps/stores/models/label.js
