+++ js/apps/stores/views/program_stores/list_item.js
@@ -4,7 +4,7 @@ define(function(require) {
         context = require('context'),
         Templates = require('handlebarsTemplates'),
         DeleteModal = require('stores/views/program_stores/delete_modal'),
-        TagsModal = require('stores/views/program_stores/tags_modal');
+        LabelsModal = require('stores/views/program_stores/labels_modal');
 
     var View = Backbone.View.extend({
         template: Templates['stores/program_stores/list_item'],
@@ -12,7 +12,7 @@ define(function(require) {
         className: 'program-store-list-item',
         events: {
             'change .status': 'update',
-            'click .view-tags': 'displayTagsModal',
+            'click .view-labels': 'displayLabelsModal',
             'click .remove': 'displayRemoveDialog'
         },
         render: function() {
@@ -72,9 +72,9 @@ define(function(require) {
             var modal = new DeleteModal({model: this.model});
             $("body").append(modal.render().el);
         },
-        displayTagsModal: function(e) {
+        displayLabelsModal: function(e) {
             e.preventDefault();
-            var modal = new TagsModal({model: this.model});
+            var modal = new LabelsModal({model: this.model});
             $('body').append(modal.render().el);
         }
     });
