+++ templates/handlebars/stores/program_stores/label_item.hbs
