+++ templates/handlebars/stores/program_stores/labels.hbs
@@ -1,0 +1,7 @@
+<div class="bbm-button cancel-button pull-right"><i class="ic ic_x"></i></div>
+<div class="labels-modal store-modal">
+    <h2>{{name}} Labels</h2>
+    <div>
+        <ul class="labels"></ul>
+    </div>
+</div>
