+++ templates/handlebars/stores/program_stores/list_item.hbs
@@ -1,5 +1,5 @@
 <td><a href="/stores/{{storeUUID}}" title="View Store">{{name}}</a></td>
-<td><a href="#" class="view-tags" data-bypass="true" title="{{labels}}">{{labels}}</a></td>
+<td><a href="#" class="view-labels" data-bypass="true" title="{{labels}}">{{labels}}</a></td>
 <td>{{address}}</td>
 <td>{{latitude}}</td>
 <td>{{longitude}}</td>
