--- templates/handlebars/stores/program_stores/tags.hbs
@@ -1,7 +1,0 @@
-<div class="bbm-button cancel-button pull-right"><i class="ic ic_x"></i></div>
-<div class="tags-modal store-modal">
-    <h2>{{name}} Tags</h2>
-    <div>
-        <ul class="tags"></ul>
-    </div>
-</div>
