+++ templates/handlebars/stores/stores/show.hbs
@@ -4,7 +4,7 @@
             <dt>Account:</dt>
             <dd>{{accountName}}</dd>
 
-            <dt>Store Tag:</dt>
+            <dt>Tag:</dt>
             <dd>{{storeTag}}</dd>
 
             <dt>Address:</dt>
