+++ templates/handlebars/stores/uploads/events/_store_data.hbs
@@ -4,7 +4,7 @@
         <ul>
             <li><strong>Row Number:</strong> {{store.row}}</li>
             <li><strong>Account:</strong> {{store.account}}</li>
-            <li><strong>Store Tag:</strong> {{store.storeTag}}</li>
+            <li><strong>Tag:</strong> {{store.storeTag}}</li>
             {{#if store.address}}
                 <li><strong>Address:</strong> {{store.address}}</li>
             {{/if}}
